// constants/enums.ts
export enum OrderStatus {
  Pending = "pending",
  Confirmed = "confirmed",
  Processing = "processing",
  PickedUp = "picked_up",
  InTransit = "in_transit",
  Delivered = "delivered",
  Cancelled = "cancelled",
  ReturnInProgress = "return_in_progress",
  ReturnedToVendor = "returned_to_vendor",
  FailedDelivery = "failed_delivery",
  Returned = "returned",
  Refunded = "refunded",
}